+++ bird-species-ui/src/components/Footer.jsx
@@ -1,5 +1,3 @@
-import { Link } from "react-router-dom";
-
 function Footer() {
   return (
     <div className="text-center p-3 bg-dark text-light mt-auto">
